+++ src/extensions/command-extension.ts
@@ -1,4 +1,5 @@
 import { toSlug } from "@furiouzz/lol/string/string";
+import type { Menu } from "obsidian";
 import type { Environment, Extension } from "../environment";
 import type { Template } from "../template";
 import { alertWrap } from "../utils/alert";
@@ -21,9 +22,10 @@ export default function (): Extension {
             env.processors.set("codeblock:command", {
                 type: "codeblock",
                 languages: { "pochoir-command": "yaml" },
-                async preprocess({ codeBlock, template }) {
+                beforePreprocess({ template }) {
                     cmd.deleteAllFromPath(template.info.file.path);
-
+                },
+                async preprocess({ codeBlock, template }) {
                     const json = tryParseYaml<
                         Partial<
                             Omit<Command, "triggers"> & {
@@ -40,15 +42,21 @@ export default function (): Extension {
                     if (triggers.length === 0) triggers.push("command");
 
                     const command: Command = {
-                        id:
-                            json?.id ??
-                            toSlug(json?.title ?? template.info.file.basename),
+                        id: json?.id ?? "",
                         title: template.info.file.basename,
                         icon: "file-question-mark",
                         action: "create",
                         ...json,
                         triggers,
                     };
+
+                    if (!command.id) {
+                        command.id = [
+                            command.action,
+                            command.triggers.join("-"),
+                            toSlug(command.title),
+                        ].join("-");
+                    }
 
                     alertWrap(() => cmd.add(template, command));
                 },
@@ -63,30 +71,49 @@ export default function (): Extension {
                     { suggestion: "icon: {icon}" },
                     { suggestion: "trigger: ribbon" },
                     { suggestion: "trigger: command" },
+                    { suggestion: "trigger: editor-menu" },
                     { suggestion: "action: insert" },
+                    { suggestion: "action: insert-from-selection" },
+                    { suggestion: "action: insert-from-clipboard" },
                     { suggestion: "action: create" },
+                    { suggestion: "action: create-from-selection" },
+                    { suggestion: "action: create-from-clipboard" },
                 ],
             });
         },
     };
 }
 
-export type CommandTrigger = "ribbon" | "command";
+export type CommandTrigger = "ribbon" | "command" | "editor-menu";
 
 export interface Command {
     id: string;
     icon: string;
     title: string;
-    action: "create" | "insert";
+    action:
+        | "create"
+        | "create-from-selection"
+        | "create-from-clipboard"
+        | "insert"
+        | "insert-from-selection"
+        | "insert-from-clipboard";
     triggers: CommandTrigger[];
 }
 
 export class CommandManager {
     env: Environment;
     map = new Map<string, string>();
+    menus = new Map<string, (menu: Menu) => void>();
 
     constructor(env: Environment) {
         this.env = env;
+        this.env.plugin.registerEvent(
+            this.env.app.workspace.on("editor-menu", (menu) => {
+                for (const cb of this.menus.values()) {
+                    cb(menu);
+                }
+            }),
+        );
     }
 
     add(template: Template, command: Command) {
@@ -97,16 +124,37 @@ export class CommandManager {
         if (triggers.includes("command")) {
             this.addCommand(template, command);
         }
+        if (triggers.includes("editor-menu")) {
+            this.addEditorMenu(template, command);
+        }
     }
 
+    performAction = async (
+        currentTemplate: Template,
+        action: Command["action"],
+    ) => {
+        let template: Template | undefined = currentTemplate;
+
+        if (action.includes("from-selection")) {
+            template = await this.env.createVirtualTemplate({
+                type: "selection",
+            });
+        } else if (action.includes("from-clipboard")) {
+            template = await this.env.createVirtualTemplate({
+                type: "clipboard",
+            });
+        }
+        if (!template) return;
+
+        if (action.startsWith("insert")) {
+            await this.env.insertFromTemplate(template);
+        } else if (action.startsWith("create")) {
+            await this.env.createFromTemplate(template);
+        }
+    };
+
     addRibon(template: Template, command: Command) {
-        const callback = async () => {
-            if (command.action === "insert") {
-                await this.env.insertFromTemplate(template);
-            } else {
-                await this.env.createFromTemplate(template);
-            }
-        };
+        const callback = () => this.performAction(template, command.action);
 
         this.env.plugin.addRibbonIcon(command.icon, command.title, callback);
         const { leftRibbon } = this.env.plugin.app.workspace;
@@ -132,22 +180,34 @@ export class CommandManager {
     }
 
     addCommand(template: Template, { id, title: name, action }: Command) {
+        const callback = () => this.performAction(template, action);
         this.env.plugin.addCommand({
             id,
             name,
-            callback: async () => {
-                if (action === "insert") {
-                    await this.env.insertFromTemplate(template);
-                } else {
-                    await this.env.createFromTemplate(template);
-                }
-            },
+            callback,
         });
         this.map.set(id, template.info.file.path);
     }
 
     removeCommand(id: string) {
         this.env.plugin.removeCommand(id);
+    }
+
+    addEditorMenu(
+        template: Template,
+        { id, title: name, action, icon }: Command,
+    ) {
+        const callback = () => this.performAction(template, action);
+        this.menus.set(id, (menu: Menu) => {
+            menu.addItem((item) => {
+                item.setTitle(name).setIcon(icon).onClick(callback);
+            });
+        });
+        this.map.set(id, template.info.file.path);
+    }
+
+    removeEditorMenu(id: string) {
+        this.menus.delete(id);
     }
 
     deleteAllFromPath(path: string) {
@@ -158,6 +218,7 @@ export class CommandManager {
         for (const [id] of entries) {
             this.removeRibon(id);
             this.removeCommand(id);
+            this.removeEditorMenu(id);
             this.map.delete(id);
         }
     }